re-wrote test script and added test data files
Tweaked how tests are identified to match test data files. It now
simply looks for a line with #N on.

Algorithm changed. It now loads the data into an array of pairs and
iterates over that. A pair will either clash (fail) or partial match, or
not match at all the current sets. On the latter we branch and try both
options of adding the pair to the current sets.